+++ gridcapa-swe-runner-app/src/main/java/com/farao_community/farao/swe/runner/app/dichotomy/DichotomyLogging.java
@@ -6,9 +6,19 @@
  */
 package com.farao_community.farao.swe.runner.app.dichotomy;
 
+import com.farao_community.farao.data.crac_api.Crac;
+import com.farao_community.farao.data.rao_result_api.RaoResult;
+import com.farao_community.farao.dichotomy.api.results.DichotomyResult;
+import com.farao_community.farao.monitoring.voltage_monitoring.VoltageMonitoringResult;
+import com.farao_community.farao.rao_runner.api.resource.RaoResponse;
 import com.farao_community.farao.swe.runner.app.configurations.DichotomyConfiguration.Parameters;
+import com.farao_community.farao.swe.runner.app.domain.SweData;
 import org.slf4j.Logger;
 import org.springframework.stereotype.Service;
+
+import java.util.Collection;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * @author Theo Pascoli {@literal <theo.pascoli at rte-france.com>}
@@ -17,6 +27,16 @@ import org.springframework.stereotype.Service;
 public class DichotomyLogging {
 
     private final Logger businessLogger;
+    private static final String SUMMARY = "Summary [{}] :  " +
+            "Limiting event : {},  \n" +
+            "Limiting element : {},  \n" +
+            "PRAs : {},  \n" +
+            "CRAs : {}.";
+    private static final String SUMMARY_BD = "Summary BD [{}] :  " +
+            "Current TTC : {},  \n" +
+            "Previous TTC : {},  \n" +
+            "Voltage Check : {},  \n" +
+            "Angle Check : {}.";
 
     public DichotomyLogging(Logger businessLogger) {
         this.businessLogger = businessLogger;
@@ -32,5 +52,29 @@ public class DichotomyLogging {
 
     public void logEndAllDichotomies() {
         businessLogger.info("All - Dichotomies are done");
+    }
+
+    public  void generateSummaryEvents(DichotomyDirection direction, DichotomyResult<RaoResponse> dichotomyResult, SweData sweData, Optional<VoltageMonitoringResult> voltageMonitoringResult) {
+        String limitingElement = "NONE";
+        String printablePrasIds = "NONE";
+        String printableCrasIds = "NONE";
+        String currentTtc = String.valueOf(dichotomyResult.getHighestValidStepValue());
+        String previousTtc = String.valueOf(dichotomyResult.getLowestInvalidStepValue());
+        String voltageCheckStatus =  voltageMonitoringResult.isPresent() ? String.valueOf(voltageMonitoringResult.get().getStatus()) : "NONE";
+        String angleCheckStatus = "NONE"; // todo after angle check monitoring for ES-PT
+        String limitingCause = dichotomyResult.getLimitingCause() != null ? DichotomyResultHelper.limitingCauseToString(dichotomyResult.getLimitingCause()) : "NONE";
+        Crac crac = (direction == DichotomyDirection.ES_FR || direction == DichotomyDirection.FR_ES) ? sweData.getCracFrEs().getCrac() : sweData.getCracEsPt().getCrac();
+        if (dichotomyResult.hasValidStep() && dichotomyResult.getHighestValidStep().getRaoResult() != null) {
+            RaoResult raoResult = dichotomyResult.getHighestValidStep().getRaoResult();
+            limitingElement = DichotomyResultHelper.getLimitingElement(crac, raoResult);
+            printablePrasIds = toString(DichotomyResultHelper.getActivatedActionInPreventive(crac, raoResult));
+            printableCrasIds = toString(DichotomyResultHelper.getActivatedActionInCurative(crac, raoResult));
+        }
+        businessLogger.info(SUMMARY, direction, limitingCause, limitingElement, printablePrasIds, printableCrasIds);
+        businessLogger.info(SUMMARY_BD, direction, currentTtc, previousTtc, voltageCheckStatus, angleCheckStatus);
+    }
+
+    private static String toString(Collection<String> c) {
+        return c.stream().map(Object::toString).collect(Collectors.joining(", "));
     }
 }
