+++ gridcapa-swe-runner-app/src/main/java/com/farao_community/farao/swe/runner/app/dichotomy/DichotomyParallelization.java
@@ -6,8 +6,6 @@
  */
 package com.farao_community.farao.swe.runner.app.dichotomy;
 
-import com.farao_community.farao.data.crac_api.Crac;
-import com.farao_community.farao.data.rao_result_api.RaoResult;
 import com.farao_community.farao.dichotomy.api.results.DichotomyResult;
 import com.farao_community.farao.monitoring.voltage_monitoring.VoltageMonitoringResult;
 import com.farao_community.farao.rao_runner.api.resource.RaoResponse;
@@ -25,9 +23,7 @@ import org.slf4j.Logger;
 import org.slf4j.MDC;
 import org.springframework.stereotype.Service;
 
-import java.util.Collection;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 /**
  * @author Theo Pascoli {@literal <theo.pascoli at rte-france.com>}
@@ -40,16 +36,6 @@ public class DichotomyParallelization {
     private final OutputService outputService;
     private final VoltageCheckService voltageCheckService;
     private final CneFileExportService cneFileExportService;
-    public static final String SUMMARY = "Summary [{}] :  " +
-            "Limiting event : {},  \n" +
-            "Limiting element : {},  \n" +
-            "PRAs : {},  \n" +
-            "CRAs : {}.";
-    public static final String SUMMARY_BD = "Summary BD [{}] :  " +
-            "Current TTC : {},  \n" +
-            "Previous TTC : {},  \n" +
-            "Voltage Check : {},  \n" +
-            "Angle Check : {}.";
 
     public DichotomyParallelization(Logger businessLogger, DichotomyLogging dichotomyLogging, DichotomyRunner dichotomyRunner,
                                     OutputService outputService, VoltageCheckService voltageCheckService,
@@ -81,12 +67,12 @@ public class DichotomyParallelization {
         // propagate in logs MDC the task requestId as an extra field to be able to send logs with calculation tasks.
         MDC.put("gridcapa-task-id", sweData.getId());
         DichotomyResult<RaoResponse> dichotomyResult = dichotomyRunner.run(sweData, direction);
+        dichotomyLogging.logEndOneDichotomy(direction);
         // Generate files specific for one direction (cne, cgm, voltage) and add them to the returned object (to create)
         String highestValidStepUrl = cneFileExportService.exportCneUrl(sweData, dichotomyResult, true, ProcessType.D2CC, direction);
         String lowestInvalidStepUrl = cneFileExportService.exportCneUrl(sweData, dichotomyResult, false, ProcessType.D2CC, direction);
         Optional<VoltageMonitoringResult> voltageMonitoringResult = voltageCheckService.runVoltageCheck(sweData, dichotomyResult, direction);
-        dichotomyLogging.logEndOneDichotomy(direction);
-        generateSummaryEvents(direction, dichotomyResult, sweData, voltageMonitoringResult);
+        dichotomyLogging.generateSummaryEvents(direction, dichotomyResult, sweData, voltageMonitoringResult);
         // fill response for one dichotomy
         return new SweDichotomyResult(direction, dichotomyResult, voltageMonitoringResult, highestValidStepUrl, lowestInvalidStepUrl);
     }
@@ -97,30 +83,5 @@ public class DichotomyParallelization {
             throw new SweInvalidDataException(String.format("No dichotomy result found for direction: [ %s ]", direction.name()));
         }
         return result.get();
-    }
-
-    private void generateSummaryEvents(DichotomyDirection direction, DichotomyResult<RaoResponse> dichotomyResult, SweData sweData, Optional<VoltageMonitoringResult> voltageMonitoringResult) {
-        String limitingElement = "NONE";
-        String printablePrasIds = "NONE";
-        String printableCrasIds = "NONE";
-        String currentTtc = String.valueOf(dichotomyResult.getHighestValidStepValue());
-        String previousTtc = String.valueOf(dichotomyResult.getLowestInvalidStepValue());
-        String voltageCheckStatus =  voltageMonitoringResult.isPresent() ? String.valueOf(voltageMonitoringResult.get().getStatus()) : "NONE";
-        String angleCheckStatus = "NONE"; // todo after angle check monitoring for ES-PT
-        String limitingCause = dichotomyResult.getLimitingCause() != null ? DichotomyResultHelper.limitingCauseToString(dichotomyResult.getLimitingCause()) : "NONE";
-        Crac crac = sweData.getCrac(direction);
-        if (dichotomyResult.hasValidStep() && dichotomyResult.getHighestValidStep().getRaoResult() != null) {
-            RaoResult raoResult = dichotomyResult.getHighestValidStep().getRaoResult();
-            limitingElement = DichotomyResultHelper.getLimitingElement(crac, raoResult);
-            printablePrasIds = toString(DichotomyResultHelper.getActivatedActionInPreventive(crac, raoResult));
-            printableCrasIds = toString(DichotomyResultHelper.getActivatedActionInCurative(crac, raoResult));
-        }
-
-        businessLogger.info(SUMMARY, direction, limitingCause, limitingElement, printablePrasIds, printableCrasIds);
-        businessLogger.info(SUMMARY_BD, direction, currentTtc, previousTtc, voltageCheckStatus, angleCheckStatus);
-    }
-
-    private static String toString(Collection<String> c) {
-        return c.stream().map(Object::toString).collect(Collectors.joining(", "));
     }
 }
