+++ gridcapa-swe-runner-app/src/main/java/com/farao_community/farao/swe/runner/app/domain/SweData.java
@@ -6,10 +6,8 @@
  */
 package com.farao_community.farao.swe.runner.app.domain;
 
-import com.farao_community.farao.data.crac_api.Crac;
 import com.farao_community.farao.data.crac_creation.creator.cim.crac_creator.CimCracCreationContext;
 import com.farao_community.farao.swe.runner.api.resource.ProcessType;
-import com.farao_community.farao.swe.runner.app.dichotomy.DichotomyDirection;
 import com.powsybl.iidm.network.Network;
 
 import java.time.OffsetDateTime;
@@ -77,11 +75,4 @@ public class SweData {
         return jsonCracPathFrEs;
     }
 
-    public Crac getCrac(DichotomyDirection direction) {
-        if (direction == DichotomyDirection.ES_FR || direction == DichotomyDirection.FR_ES) {
-            return cracFrEs.getCrac();
-        } else {
-            return cracEsPt.getCrac();
-        }
-    }
 }
